requirements.txt
altgraph==0.17.4
boto3==1.37.21
botocore==1.37.21
colorama==0.4.6
iniconfig==2.1.0
jmespath==1.0.1
packaging==24.2
pefile==2023.2.7
pluggy==1.5.0
pycryptodome==3.22.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
s3transfer==0.11.4
secret==0.8
setuptools==78.1.0
six==1.17.0
tabulate==0.9.0
urllib3==2.3.0
